Login/Signup Pages: Add bg images
Replaced bg-sphere with bg-graph as per the design repo.

Quick fix to display the bg images for signup and login
The image is slightly displaced from the original design and
sticking to the design would require more changes in the
application.html.

Part of: https://github.com/AICrowd/AIcrowd-Design/issues/62