Travis: Fix challenges test
The test go to "/" then try to click on the link to
a challenge. Only 3 challenges are shown on the main
pages, which made me think that the challenge being tested
may not appear/may not be clickable.
This commit modifies the tests to go to "/challenges" before
clicking on the name of the challenge.