Leaderboard Calculation: Fix Bug
When we select the submissions we need to consider Team Participants
only from teams of the current round. If we do not do this, and a
participant is part of 2 teams across 2 challegens, the leaderboard
computation will fail, as submissions by this user will be selected
twice and error out during adding this to the temp_stats table,
which requires uniq values only.