Teams: Add support for max_team_participants and teams_allowed
We should be able to mark challenges where teams are allowed and
where teams are not. This should allow for no teams to be created.
Also started respecting the 'max participants per team' variable.
This is also used to show the number of invitations that can be
sent by the team organizer.