Challenges: Revive support for ordering Featured Challenges
This commit, add a new page for reordering the challenges
The top 6 are shown on the landing page. Changed the default ordering
from desc and modified the way the seq is set.
Every new challenge will be at the bottom of the list.
Also optimized the JS that runs on each page of the challenges controller.
Removed excess sorttable module as there is already one in jquery.

Closes https://github.com/AIcrowd/AIcrowd/issues/777