Notification Settings: Fixup to allow changes
The data related to newsletter and other notifications setting
was saved in a seperate model and table called 'Email Notifications'
This is of no use, and should be removed in upcoming refactors.
This commit decouples the model from notification settings page
and allows us to use the form directly to change newsletter settings.